% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_wind_power.R
\name{compute_wind_power}
\alias{compute_wind_power}
\title{Compute Wind Power}
\usage{
compute_wind_power(blade_length, wind_speed, CP = 0.45, pi = 22/7, mass = 2360)
}
\arguments{
\item{blade_length}{in m}

\item{pi}{is 22/7}

\item{coefficient_of_performance}{is the efficiency of the windmill (percent). default is 0.45}

\item{air}{density (kg/m^3) default is 2360}

\item{windspeed}{is the speed of the wind blowing in m/s}
}
\description{
power of wind turbine
This function computes power in watts and kinetic energy in Joules
generated by a wind turbine given its blade length and wind speeds
}
\examples{
compute_wind_power(20,30)
}
\author{
Nadine, Adele, and Hannah
}
