% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbines_wind.R
\docType{data}
\name{turbines_wind}
\alias{turbines_wind}
\title{Speeds and wind turbine blade lengths}
\format{
A data frame with 100 rows and 2 variable
\itemize{
\item speeds (m/s)
\item blade_lengths (meters)
}
}
\source{
\url{colwell_snyder_irish.com}
}
\usage{
turbines_wind
}
\description{
Data from from Colwell, Snyder, and Irish, LLC
}
\keyword{datasets}
